Split SftpClient implementation over several files.
