+++ src/Tmds.Ssh/SftpException.cs
@@ -1,0 +1,12 @@
+namespace Tmds.Ssh
+{
+    public class SftpException : SshOperationException
+    {
+        public SftpError Error { get; private set; }
+
+        internal SftpException(SftpError error) : base(error.ToString())
+        {
+            Error = error;
+        }
+    }
+}
